Start of the client.  Loading and expanding data from the server.
The ResourceService is the heavy lifter here.  It will cache objects
using the ResourceCacheService.  If the resource isn't there it will see
if it already has a promise for the resource, if so it returns that
promise.  If no promise is pending it will make the request.

When it gets a response it will set the resource in cache and then
load the resource's relationships.