Wait for all nested objects to be loaded before resolving promise
This was cuasing some weird behavior where a promise for productions
would resolve, but its products were undefined when accessed because
they were still pending.  This change has it wait for all promises to
get its relationships.

Future upgrade is to collect the needed ids and do a single call in
order to cut down on api calls to the server.