let GoogleChartService format data and wait to draw
The service can now take an array of data and attributes or functions
for hwo to get certain data out of each object and it will build the
array of chart data ready to draw.

The service gains a draw function which will wait for its own promise of
being loaded so other services can just call draw and not know that
it needs to load.