Allow choosing a date range to view
Introduce new element to work with a stats widget to provide date range
selections.  New DateRangeWidget base class also helps initialize
and observe attributes and calls for new data when they are updated.

There is still an issue rendering a chart when there are no records in
the date range.  No records means it can't get any column titles and it
can't make a chart with one column.  Maybe create a fake record to
populate?