New widget to enter or update productions
Moving through products or dates will lookup an existing production for
that combination so its count can be updated.

Saving a production for a date before today will automatically move the
date ahead a day to assist entering old data.  Maybe this should switch
to cycling through products in the future?